listening-comps/requirements.txt
youtube-transcript-api>=0.6.2
chromadb>=0.4.22
google-generativeai>=0.8.3
python-dotenv>=1.0.1
openai>=1.0.1
transformers>=4.37.0
accelerate
torch>=2.2.0
scipy>=1.12.0
ffmpeg-python>=0.2.0
streamlit
